requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.20.0
tqdm>=4.60.0
matplotlib>=3.5.0
pandas>=1.4.0
transformers>=4.20.0
scikit-learn>=1.0.0
scipy>=1.7.0
pyyaml>=6.0
pillow>=9.0.0

# Code generation dependencies
tokenizers>=0.12.0
tree-sitter>=0.20.0
pygments>=2.10.0
astroid>=2.8.0
yapf>=0.32.0
autopep8>=1.6.0

# API dependencies
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.8.0
httpx>=0.20.0
jsonschema>=4.0.0

# Testing and development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=23.0.0
isort>=5.10.0
mypy>=0.900
pylint>=2.12.0
coverage>=6.0.0
tox>=3.24.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.12.0

# Jupyter notebooks
ipywidgets>=8.0.0
jupyterlab>=3.2.0
ipykernel>=6.4.0
